Refactoring `ParticipantsTable` for upcoming Modal
